Long에서 int

+++ src/main/java/com/izo/itaportal/model/User.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class User {
 
-    private Long idUser;
+    private int idUser;
     private String logId;
     private String password;
     private String role;
